% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{CleanAmericanStates}
\alias{CleanAmericanStates}
\title{Clean states from the USA}
\usage{
CleanAmericanStates(states)
}
\arguments{
\item{state}{A vector of state names}
}
\value{
A vector of cleaned state names
}
\description{
Clean states from the USA
}
