% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{CleanCanadaStates}
\alias{CleanCanadaStates}
\title{Clean states from Canada}
\usage{
CleanCanadaStates(states)
}
\arguments{
\item{state}{A vector of state names}
}
\value{
A vector of cleaned state names
}
\description{
Clean states from Canada
}
