% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{CleanIndianStates}
\alias{CleanIndianStates}
\title{Clean Indian States}
\usage{
CleanIndianStates(states)
}
\arguments{
\item{state}{A vector of state names}
}
\value{
A vector of cleaned state names
}
\description{
Clean Indian States
}
