% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{CleanSouthAfricanStates}
\alias{CleanSouthAfricanStates}
\title{Clean states from South Africa}
\usage{
CleanSouthAfricanStates(states)
}
\arguments{
\item{state}{A vector of state names}
}
\value{
A vector of cleaned state names
}
\description{
Clean states from South Africa
}
