% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{CollapseLineageToVOCs}
\alias{CollapseLineageToVOCs}
\title{Collpase pangolin lineage a list of VOCs}
\usage{
CollapseLineageToVOCs(variant_df, vocs = GetVOCs(), custom_voc_mapping = NULL)
}
\arguments{
\item{variant_df}{A dataframe with "pangolin_lineage" column}

\item{vocs}{A named list with VOC (variant of concernt) name as key and a list of lineages under the VOC.
The method will replace all lineages under a VOC to its name.}

\item{custom_voc_mapping}{A named vector with a custom mapping for naming some lineages. See example.}
}
\value{
A dataframe with a new column "lineage_collpased".
}
\description{
Collpase pangolin lineage a list of VOCs
}
