% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{CombineCases}
\alias{CombineCases}
\title{Create a combined dataframe of sequenced cases and confirmed cases}
\usage{
CombineCases(cases_sequenced, confirmed_long)
}
\arguments{
\item{cases_sequenced}{A long dataframe of per state sequenced cases}

\item{cases_total}{A long dataframe of total monthly cases}
}
\value{
a combined dataframe with case load and sequenced
}
\description{
Create a combined dataframe of sequenced cases and confirmed cases
}
