% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{CountsToPrevalence}
\alias{CountsToPrevalence}
\title{Convert monthwise counts to prevalence}
\usage{
CountsToPrevalence(variant_df)
}
\arguments{
\item{variant_df}{A dataframe}
}
\value{
A dataframe with monthwise prevalence of variants
}
\description{
Convert monthwise counts to prevalence
}
