% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{FilterGISAIDIndia}
\alias{FilterGISAIDIndia}
\title{Filter GISAID India Metadata for India}
\usage{
FilterGISAIDIndia(gisaid_metadata_all)
}
\arguments{
\item{gisaid_metadata_all}{A dataframe with all GISAID metadata}
}
\value{
A dataframe with only Indian entries in Human and where the date is known
}
\description{
Filter GISAID India Metadata for India
}
