% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gisaid.R
\name{FormatGISAIDMetadata}
\alias{FormatGISAIDMetadata}
\title{Format metadata.tar.x file downloaded from GISAID with some missing columns}
\usage{
FormatGISAIDMetadata(
  df,
  collection_col = "Collection date",
  submission_col = "Submission date"
)
}
\arguments{
\item{df}{GISAID metadata dataframe}

\item{collection_col}{Name of column representing date of collection (Collection date)}

\item{submission_col}{Name of column representing date of submission (Submission date)}
}
\value{
A data frame with following extra columns:
DateCollected, MonthYearCollected, YearCollected, MonthCollected, WeekCollected
DateCollectedNumeric, MonthYearCollectedNumeric
DateSubmitted, MonthYearSubmitted, YearSubmitted, MonthSubmitted, WeekSubmitted
DateSubmittedNumeric, MonthYearSubmittedNumeric
}
\description{
Format metadata.tar.x file downloaded from GISAID with some missing columns
}
