% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode.R
\name{GeocodeAddress}
\alias{GeocodeAddress}
\title{Geocode a given address to get its latitude, longitude from Gooogle}
\usage{
GeocodeAddress(address, api.key = NULL, restrict.to = NULL)
}
\arguments{
\item{address}{String address to geocode}

\item{api.key}{API key for Google Maps}

\item{restrict.to}{Country name where the search should be restricted to,
Currently only supports 'India'}
}
\value{
A dataframe with geocoded address
}
\description{
Geocode a given address to get its latitude, longitude from Gooogle
}
