% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_data.R
\name{GetIndiaConfirmedCases}
\alias{GetIndiaConfirmedCases}
\title{Get daily confirmed cases for India}
\usage{
GetIndiaConfirmedCases(
  url = "https://data.covid19bharat.org/csv/latest/state_wise_daily.csv"
)
}
\arguments{
\item{url}{URL to fetch data from}
}
\value{
A data frame containing daily cases for each state
}
\description{
Get daily confirmed cases for India
}
