% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_data.R
\name{GetIndiaConfirmedCasesWeeklyLong}
\alias{GetIndiaConfirmedCasesWeeklyLong}
\title{Get India weekwise cases long}
\usage{
GetIndiaConfirmedCasesWeeklyLong(
  url = "https://data.covid19bharat.org/csv/latest/state_wise_daily.csv",
  level = "State"
)
}
\arguments{
\item{url}{URL to fetch data from}
}
\value{
A data frame containing monthly cases for each state in long form
}
\description{
Get India weekwise cases long
}
