% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_data.R
\name{GetIndiaDailyCasesCumulative}
\alias{GetIndiaDailyCasesCumulative}
\title{Get daily cases for India for a given status type}
\usage{
GetIndiaDailyCasesCumulative(
  url = "https://data.covid19bharat.org/csv/latest/state_wise_daily.csv",
  status = "Confirmed"
)
}
\arguments{
\item{url}{URL to fetch data from}

\item{status}{One of "Confirmed", "Recovered", or "Deceased"}
}
\value{
A data frame containing daily cases for each state
}
\description{
Get daily cases for India for a given status type
}
