% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_data.R
\name{GetIndiaDailyData}
\alias{GetIndiaDailyData}
\title{Get daily statewise data for India}
\usage{
GetIndiaDailyData(
  url = "https://saketkc.github.io/covid19bharat_data//csv/latest/state_wise_daily.csv"
)
}
\arguments{
\item{url}{URL to fetch data from}
}
\value{
A data frame containing daily cases for each state
}
\description{
The 'Status' column is one of 'Confirmed', 'Deceased' or 'Recovered'
}
