% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_data.R
\name{GetIndiaDeceasedCasesCumulative}
\alias{GetIndiaDeceasedCasesCumulative}
\title{Get daily deceased cases for India}
\usage{
GetIndiaDeceasedCasesCumulative(
  url = "https://data.covid19bharat.org/csv/latest/state_wise_daily.csv"
)
}
\arguments{
\item{url}{URL to fetch data from}
}
\value{
A data frame containing daily cases for each state
}
\description{
Get daily deceased cases for India
}
