% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_data.R
\name{GetIndianStates}
\alias{GetIndianStates}
\title{Get a list of Indian name states}
\usage{
GetIndianStates()
}
\value{
A list with state abbreviation as the key and full name as the value
}
\description{
Get a list of Indian name states
}
