% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{GetMonthYear}
\alias{GetMonthYear}
\title{Get MonthYear (zoo) from a dataframe string formatted data column}
\usage{
GetMonthYear(datecol, datefmt = "\%Y-\%m-\%d")
}
\arguments{
\item{datecol}{A vector of string formatted dates}

\item{datefmt}{Format of the date (default is Y-m-d)}
}
\value{
A vector with dates converted to MonthYear format (zoo::as.yearmon)
}
\description{
Get MonthYear (zoo) from a dataframe string formatted data column
}
