% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{PlotSequencedPropHeatmap}
\alias{PlotSequencedPropHeatmap}
\title{Plot heatmap of cases sequenced}
\usage{
PlotSequencedPropHeatmap(cases_and_shared)
}
\arguments{
\item{cases_and_shared}{A dataframe containing counts of confirmed cases and sequenced cases}
}
\value{
A heatmap of total percentage of cases sequenced
}
\description{
Plot heatmap of cases sequenced
}
