% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{PlotTotalHeatmap}
\alias{PlotTotalHeatmap}
\title{Plot heatmap of total cases/sequences}
\usage{
PlotTotalHeatmap(df, color_legend = "Total cases")
}
\arguments{
\item{cases_and_shared}{A dataframe containing counts of confirmed cases or sequenced cases (any column
named 'value')}
}
\value{
A heatmap of total cases/sequences
}
\description{
Plot heatmap of total cases/sequences
}
