% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gisaid.R
\name{ReadAuspiceInstrument}
\alias{ReadAuspiceInstrument}
\title{Read instrument metadata from GISAID batch downloads}
\usage{
ReadAuspiceInstrument(path)
}
\arguments{
\item{path}{Path to a list of directories with "sequence.tsv" inside each subdirector}
}
\value{
A dataframe with all the instrument related metadata
}
\description{
Read instrument metadata from GISAID batch downloads
}
