% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gisaid.R
\name{ReadAuspiceMetadata}
\alias{ReadAuspiceMetadata}
\title{Read all metadata from GISAID batch downloads}
\usage{
ReadAuspiceMetadata(path)
}
\arguments{
\item{path}{Path to a list of directories with "sequence.tsv" inside each subdirector}
}
\value{
A dataframe with all the collated metadata
}
\description{
Read all metadata from GISAID batch downloads
}
