% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gisaid.R
\name{ReadGISAIDMetada}
\alias{ReadGISAIDMetada}
\title{Read metadata downloaded from GISAID}
\usage{
ReadGISAIDMetada(path, showProgress = FALSE, ...)
}
\arguments{
\item{path}{Path to tar.xz or metadata.txt downloaded from GISAID}

\item{showProgress}{Whether to print progress bars for fread}

\item{...}{Other parameters for data.table::fread}
}
\value{
A data frame with parsed metadata with some extra columns
}
\description{
Read metadata downloaded from GISAID
}
