% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{SanitizeColumn}
\alias{SanitizeColumn}
\title{Sanitize a vector to remove special characters and convert it to tile}
\usage{
SanitizeColumn(col)
}
\arguments{
\item{col}{String vector}
}
\value{
A string vector with sanitized entries
}
\description{
This function removes:
1. Commas with space
2. Hyphen with space
3. Multiple dots with single space
4. Multiple spacies with single space
5. NA with "UNKNOWN"
}
