% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{SummarizeVariantsMonthwise}
\alias{SummarizeVariantsMonthwise}
\title{Summarize the total number of variants per month}
\usage{
SummarizeVariantsMonthwise(variant_df)
}
\arguments{
\item{variant_df}{A dataframe}
}
\value{
A dataframe with monthwise counts of each variant sequenced
}
\description{
Summarize the total number of variants per month
}
