% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{SummarizeVariantsWeekwise}
\alias{SummarizeVariantsWeekwise}
\title{Summarize the total number of variants per week}
\usage{
SummarizeVariantsWeekwise(variant_df, by_state = FALSE)
}
\arguments{
\item{variant_df}{A dataframe}
}
\value{
A dataframe with monthwise counts of each variant sequenced
}
\description{
Summarize the total number of variants per week
}
