% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{TotalSequencesPerMonthCountrywise}
\alias{TotalSequencesPerMonthCountrywise}
\title{Get total sequenced samples per country}
\usage{
TotalSequencesPerMonthCountrywise(variant_df, rename_country_as_state = TRUE)
}
\arguments{
\item{variant_df}{A dataframe}
}
\value{
A dataframe with sequencing statistics per country
}
\description{
Get total sequenced samples per country
}
