% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{TotalSequencesPerMonthStatewise}
\alias{TotalSequencesPerMonthStatewise}
\title{Get total number of sequenced samples per month in a Country}
\usage{
TotalSequencesPerMonthStatewise(variant_df, drop_country = FALSE)
}
\value{
A dataframe with sequencing statistics per state per country
}
\description{
Get total number of sequenced samples per month in a Country
}
